+++ common/factorization/table/eulers_totient.h
@@ -1,6 +1,8 @@
 #pragma once
 
+#include "common/base.h"
 #include "common/factorization/table/primes.h"
+
 #include <vector>
 
 namespace factorization {
@@ -14,7 +16,7 @@ class EulersTotient : public Primes {
     eulers_totient.resize(Primes::table_size + 1);
     eulers_totient[0] = 0;
     eulers_totient[1] = 1;
-    for (unsigned i = 2; i <= Primes::table_size; ++i) {
+    for (uint64_t i = 2; i <= Primes::table_size; ++i) {
       unsigned p = Primes::table[i];
       eulers_totient[i] =
           eulers_totient[i / p] * (Primes::table[i / p] == p ? p : (p - 1));
