+++ common/factorization/utils/divisors.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include "common/factorization/base.h"
+
 #include <functional>
 
 inline std::vector<uint64_t> GetDivisors(const TFactorization& factorization) {
   std::vector<uint64_t> v;
-  std::function<void(unsigned, uint64_t)> AddR = [&](unsigned fpi,
-                                                     uint64_t current) -> void {
+  std::function<void(size_t, uint64_t)> AddR = [&](size_t fpi,
+                                                   uint64_t current) -> void {
     if (fpi < factorization.size()) {
       for (unsigned i = 0; i <= factorization[fpi].power; ++i) {
         AddR(fpi + 1, current);
